Developer's Docs

PeriodicOrbits.minimal_period — Function

minimal_period(ds::DynamicalSystem, po::PeriodicOrbit; kw...) → minT_po

Compute the minimal period of the periodic orbit po of the dynamical system ds. Return the periodic orbit minT_po with the minimal period. In the literature, minimal period is also called prime, principal or fundamental period.

Keyword arguments

atol = 1e-4 : After stepping the point u0 for a time T, it must return to atol neighborhood of itself to be considered periodic.

maxiter = 40 : Maximum number of Poincare map iterations. Continuous-time systems only. If the number of Poincare map iterations exceeds maxiter, but the point u0 has not returned to atol neighborhood of itself, the original period po.T is returned.

Δt = missing : The time step between points in the trajectory minT_po.points. If Δt is missing, then Δt=minT_po.T/100 is used. Continuous-time systems only.

Description

For discrete systems, a valid period would be any natural multiple of the minimal period. Hence, all natural divisors of the period po.T are checked as a potential period. A point u0 of the periodic orbit po is iterated n times and if the distance between the initial point u0 and the final point is less than atol, the period of the orbit is n.

For continuous systems, a point u0 of the periodic orbit is integrated for a very short time. The resulting point u1 is used to create a normal vector a=(u1-u0) to a hyperplane perpendicular to the trajectory at u0. A Poincare map is created using this hyperplane. Using the Poincare map, the hyperplane crossings are checked. Time of the first crossing that is within atol distance of the initial point u0 is the minimal period. At most maxiter crossings are checked.

PeriodicOrbits.isstable — Function

isstable(ds::CoreDynamicalSystem, po [, jac]) → new_po

Determine the local stability of the periodic orbit po using the jacobian rule jac. Returns a new periodic orbit for which po.stable is set to true if the periodic orbit is stable or false if it is unstable.

For discrete-time systems, the stability is determined using eigenvalues of the jacobian of po.T-th iterate of the dynamical system ds at the point po.points[1]. If the maximum absolute value of the eigenvalues is less than 1, the periodic orbit is marked as stable.

For continuous-time systems, the stability is determined by the Floquet multipliers of the monodromy matrix. If the maximum absolute value of the Floquet multipliers is less than 1 (while neglecting the multiplier which is always 1), the periodic orbit is marked as stable.

The default value of jacobian rule jac is obtained via automatic differentiation.

PeriodicOrbits.uniquepos — Function

uniquepos(pos::Vector{<:PeriodicOrbit}; atol=1e-6) → Vector{PeriodicOrbit}

Return a vector of unique periodic orbits from the vector pos of periodic orbits. By unique we mean that the distance between any two periodic orbits in the vector is greater than atol. To see details about the distance function, see podistance.

Keyword arguments

atol : minimal distance between two periodic orbits for them to be considered unique.

PeriodicOrbits.poequal — Function

poequal(po1::PeriodicOrbit, po2::PeriodicOrbit; kwargs...) → true/false

Return true if the periodic orbits po1 and po2 are equal within the given thresholds.

Keyword arguments

Tthres=1e-3 : difference in periods of the periodic orbits must be less than this threshold

dthres=1e-3 : distance between periodic orbits must be less than this threshold

distance : distance function used to compute the distance between the periodic orbits

Distance between the orbits is computed using the given distance function distance. The default distance function is StrictlyMinimumDistance(true, Euclidean()) which finds the minimal Euclidean distance between any pair of points where one point belongs to po1 and the other to po2. For other options of the distance function, see StateSpaceSets.set_distance. Custom distance function can be provided as well.

DynamicalSystemsBase.isdiscretetime — Function

isdiscretetime(po::PeriodicOrbit) → true/false

Return true if the periodic orbit belongs to a discrete-time dynamical system, false if it belongs to a continuous-time dynamical system.

isdiscretetime(ds::DynamicalSystem) → true/false

Return true if ds operates in discrete time, or false if it is in continuous time. This is information deduced from the type of ds.

PeriodicOrbits.podistance — Function

podistance(po1::PeriodicOrbit, po2::PeriodicOrbit, [, distance]) → Real

Compute the distance between two periodic orbits po1 and po2. Periodic orbits po1 and po2 and the dynamical system ds all have to be either discrete-time or continuous-time. Distance between the periodic orbits is computed using the given distance function distance. The default distance function is StrictlyMinimumDistance(true, Euclidean()) which finds the minimal Euclidean distance between any pair of points where one point belongs to po1 and the other to po2. For other options of the distance function, see StateSpaceSets.set_distance. Custom distance function can be provided as well.

PeriodicOrbits.PeriodicOrbitFinder — Type

PeriodicOrbitFinder

Supertype for all the periodic orbit detection algorithms. Each of the concrete subtypes of PeriodicOrbitFinder should represent one given algorithm for detecting periodic orbits. This subtype will include all the necessary parameters for the algorithm to work and optionally their default values.