The Public API

PeriodicOrbits.PeriodicOrbit — Type

A structure that contains information about a periodic orbit.

points::StateSpaceSet - points of the periodic orbit. This container always holds the complete orbit.

T::Real - the period of the orbit

stable::Union{Bool, Missing} - local stability of the periodic orbit. Unknown stability is set to missing.

PeriodicOrbits.InitialGuess — Type

A structure that contains an initial guess for a periodic orbit detection algorithms.

u0::AbstractArray{<:Real} - guess of a point in the periodic orbit

T::Union{Real, Nothing} - guess of period of the orbit. Some algorithms do not require the period guess to be given, in which case T is set to nothing.

PeriodicOrbits.periodic_orbit — Function

periodic_orbit(ds::DynamicalSystem, alg::PeriodicOrbitFinder, ig::InitialGuess = InitialGuess(ds)) → PeriodicOrbit

Try to find single periodic orbit of the dynamical system ds using the algorithm alg given some initial guess ig. For more details on the periodic orbit detection algorithms, see the documentation of the specific algorithm.

PeriodicOrbits.periodic_orbits — Function

periodic_orbit(ds::DynamicalSystem, alg::PeriodicOrbitFinder, igs::Vector{InitialGuess} = InitialGuess(ds)) → Vector{PeriodicOrbit}

Try to find multiple periodic orbits of the dynamical system ds using the algorithm alg given some initial guesses igs. For more details on the periodic orbit detection algorithms, see the documentation of the specific algorithm.

Algorithms for Discrete-Time Systems

SchmelcherDiakonos

DavidchackLai

Algorithms for Continuous-Time Systems

OptimizedShooting